Unify GGUF generation CLI across models
Replace the model-specialized generate_minimax_gguf entrypoint with a
generic src/cli/generate_gguf.py that dispatches by general.architecture
through the MoE registry (detect_spec). The model-agnostic greedy decode /
seed-file / dist driver now lives in src/runtime/generation.py
(run_gguf_generation + GGUFTokenRuntime), and architectures plug in via a
new MoEModelSpec.build_token_runtime hook implemented by MiniMaxM2Spec
(deferred imports keep the spec cheap for inspection/registry use).

- delete src/cli/generate_minimax_gguf.py and
  src/models/minimax_m2/gguf_generation.py
- add --architecture override (default auto-detect)
- prints gguf_load_start banner with detected architecture

Behavior unchanged: TP4 smoke via the unified CLI still yields
generated_token_ids=16474,9 at 15.853 GiB/rank.
